+++ lib/video_player_sharara.dart
@@ -1,4 +1,6 @@
 library sharara_video_player;
 
-export 'src/controller.dart';
+export 'src/controller.dart' show ShararaVideoPlayerController,UiWorker,PlayPause,ViewModeDimensions,
+Rebuilder;
+export 'src/options.dart';
 export 'src/video_player_sharara.dart';