+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+* added new implementations when video player failed to retrieve data 
+* construct all options of sharara video player inside Options class 
+* added refreshing and rebuilding internal player controller to refresh ui and refresh internal player
+
 ## 1.0.14
 
 * Update to the last version for better performance and options